Update [1.0.2]
-Over 60 User have reported unsupported devices thanks to all
-Samsung Galaxy S3 Verizon (d2vzw) added
-Device recognition issues on Samsung devices fixed
-New stuff added to support Sony and more Samsung devices (Support on
concrete devices will following with next updates)

+++ src/de/mkrtchyan/recoverytools/FlashUtil.java
@@ -9,8 +9,7 @@ public class FlashUtil {
 	
 	Context context;
 	CommonUtil cu;
-	NotificationUtil nu;
-	Support s;
+	Support s = new Support();
 	private static File fflash;
 	private static File fdump;
 	
@@ -22,8 +21,6 @@ public class FlashUtil {
 	public FlashUtil(Context context){
 		context = this.context;
 		cu = new CommonUtil(this.context);
-		nu = new NotificationUtil(this.context);
-		s = new Support();
 		if (s.MTD){
 			fflash = new File(context.getFilesDir(), "flash_image");
 			fdump = new File(context.getFilesDir(), "dump_image");
@@ -35,29 +32,45 @@ public class FlashUtil {
 	}
 	
 	public void flash(File file) {
+			
 		if (file.exists()) {
-			if (!s.MTD){
-				cu.executeShell("dd if=" + file.getAbsolutePath() + " of=" + s.RecoveryPath);
-				nu.createDialog(R.string.info, R.string.flashed, true, true);
-			} else {
+			
+			if (s.MTD){
 				cu.executeShell(fflash.getAbsolutePath() + " recovery " + file.getAbsolutePath());
+			}
+//			if (s.BLM){
+//				File tmp = new File(PathToRecoveryTools, "tmp");
+//				cu.checkFolder(tmp);
+//				
+//				cu.unzip(file, tmp);
+				
+//				File meta = new File(tmp, "META-INF");
+//				if (meta.exists())
+//					cu.deleteFolder(meta, true);
+				
+//				cu.copy(tmp, new File(s.RecoveryPath), true);
+//			}
+			if (!s.MTD
+//					&& !s.BLM
+					&& !s.RecoveryPath.equals("")){
+				cu.executeShell("dd if=" + file.getAbsolutePath() + " of=" + s.RecoveryPath);
 			}
 		}
 	}
 	
 	public void backup() {
 		if (!s.MTD){
-			nu.createDialog(R.string.bakreport, cu.executeShell("dd if=" + s.RecoveryPath + " of=" + PathToBackup.getAbsolutePath() + "/backup.img"), true);
+			cu.executeShell("dd if=" + s.RecoveryPath + " of=" + PathToBackup.getAbsolutePath() + "/backup.img");
 		} else {
-			nu.createDialog(R.string.bakreport, cu.executeShell(fdump.getAbsolutePath() + " recovery " + fBACKUP.getAbsolutePath()), true);
+			cu.executeShell(fdump.getAbsolutePath() + " recovery " + fBACKUP.getAbsolutePath());
 		}
 	}
 	
 	public void restore() {
 		if (!s.MTD){
-			nu.createDialog(R.string.resreport, cu.executeShell("dd if=" + PathToBackup.getAbsolutePath() + "/backup.img of=" + s.RecoveryPath), true);
+			cu.executeShell("dd if=" + PathToBackup.getAbsolutePath() + "/backup.img of=" + s.RecoveryPath);
 		} else {
-			nu.createDialog(R.string.resreport, cu.executeShell(fflash.getAbsolutePath() + " recovery " + fBACKUP.getAbsolutePath()), true);
+			cu.executeShell(fflash.getAbsolutePath() + " recovery " + fBACKUP.getAbsolutePath());
 		}
 	}
